Tab bar controller is now custom class
