+++ GridRace/GridRace/controllers/StartScreenViewController.swift
@@ -12,7 +12,6 @@ class StartScreenViewController: UIViewController {
     
     var startButton = UIButton()
     var backgroundImage = UIImageView()
-    let timerView = TimerView()
     
     init() {
         super.init(nibName: nil, bundle: nil)
@@ -73,45 +72,11 @@ class StartScreenViewController: UIViewController {
         //add date object to user defaults
         UserDefaults.standard.set(Date(), forKey: "FirstLaunchDate")
         
-        _ = Timer.scheduledTimer(withTimeInterval: 1, repeats: true, block: {_ in
-            let interval = Date().timeIntervalSince(AppResources.firstLaunchDate)
-            
-            let ti = NSInteger(interval)
-            
-            let seconds = ti % 60
-            let minutes = (ti / 60) % 60
-            let hours = ti / 3600
-            
-            AppResources.timeToDisplay = NSString(format: "%0.2d:%0.2d:%0.2d", hours, minutes, seconds) as String
-            self.timerView.timeLabel.text = AppResources.timeToDisplay
-        })
-        
-        let placesViewController = ObjectiveTableViewController(title: "Places", tabBarImage: #imageLiteral(resourceName: "directional_arrow"), dataCategory: ObjectiveCategory(rawValue: "places"), objectives: &AppResources.placesObjectives, data: &AppResources.placesUserData)
-        let bonusViewController = ObjectiveTableViewController(title: "Bonus", tabBarImage: #imageLiteral(resourceName: "clock_outline"), dataCategory: .bonus, objectives: &AppResources.bonusObjectives, data: &AppResources.bonusUserData)
-        
-        let mainTabController = UITabBarController()
-        
-        mainTabController.tabBar.barTintColor = AppColors.backgroundColor
-        mainTabController.tabBar.tintColor = AppColors.textPrimaryColor
-        
-        mainTabController.viewControllers = [UINavigationController(rootViewController: placesViewController), UINavigationController(rootViewController: bonusViewController)]
-        
-        addTimerToTabBar(tabBar: mainTabController.tabBar)
-        
-        present(mainTabController, animated: true, completion: nil)
+        present(CustomTabBarController(), animated: true, completion: nil)
     }
     
     func addTimerToTabBar(tabBar: UITabBar) {
-        tabBar.addSubview(timerView)
-        
-        timerView.translatesAutoresizingMaskIntoConstraints = false
-        
-        NSLayoutConstraint.activate([
-            timerView.centerXAnchor.constraint (equalTo: tabBar.centerXAnchor),
-            timerView.bottomAnchor.constraint(equalTo: tabBar.bottomAnchor),
-            timerView.heightAnchor.constraint(equalToConstant: tabBar.bounds.height),
-            timerView.widthAnchor.constraint(equalToConstant: tabBar.frame.width * 0.3)
-            ])
+       
     }
 
     override func didReceiveMemoryWarning() {
